ML 22 new values (#401)
* ML-22-initial value interface

* ML-22-engine uses df_value

* ML-22-value updates in engine and highdim

* ML-22-bool and continous values working

* ML-22-associated nan columns

* ML-22-datetime value

* ML-22-small-refactor

* ML-22-updated-docstrings

* ML-22-bugfixes

* ML-22-fixes-from-rebase

* ML-22-remove further references to domain

* ML-22-change df_value and assoc value to mappings

* ML-22-redo Nanvalues, linting

* ML-22-remove nan name

* ML-22-check-if-transformer-is-fit

* ML-22-added IntegerBool to value builder

+++ synthesized/transformer/quantile.py
@@ -47,7 +47,7 @@ class QuantileTransformer(Transformer):
         if nonnegative and not positive:
             df[self.name] = np.maximum(df[self.name], 0.001)
 
-        df[self.name] = self._transformer.transform(df[[self.name]])
+        df[self.name] = self._transformer.transform(df[[self.name]]).astype(np.float32)
 
         return df
